DarkSide_challenge/requirements.txt
pytorch==2.1.0
numpy==1.24.3
matplotlib==3.7.2
scipy==1.10.1
xarray==2022.11.0
tqdm==4.66.2